+++ Toolchain/BuildClang.sh
@@ -247,14 +247,20 @@ popd
 # === COPY HEADERS ===
 
 SRC_ROOT=$($REALPATH "$DIR"/..)
-FILES=$(find "$SRC_ROOT"/Kernel/API "$SRC_ROOT"/Userland/Libraries/LibC -name '*.h' -print)
-
+FILES=$(find \
+    "$SRC_ROOT"/Kernel/API \
+    "$SRC_ROOT"/Userland/Libraries/LibC \
+    "$SRC_ROOT"/Userland/Libraries/LibRegex/RegexDefs.h \
+    -name '*.h' -print)
 for arch in $ARCHS; do
     mkdir -p "$BUILD/${arch}clang"
     pushd "$BUILD/${arch}clang"
         mkdir -p Root/usr/include/
         for header in $FILES; do
-            target=$(echo "$header" | "$SED" -e "s|$SRC_ROOT/Userland/Libraries/LibC||" -e "s|$SRC_ROOT/Kernel/|Kernel/|")
+            target=$(echo "$header" | "$SED" \
+                -e "s|$SRC_ROOT/Userland/Libraries/LibC||" \
+                -e "s|$SRC_ROOT/Kernel/|Kernel/|" \
+                -e "s|$SRC_ROOT/Userland/Libraries/LibRegex/|LibRegex/|")
             buildstep "system_headers" "$INSTALL" -D "$header" "Root/usr/include/$target"
         done
     popd
